Added full support for GitHub repos, better script runner, and plent of small issues while I was on it.

+++ lib/shelly.rb
@@ -7,16 +7,9 @@ require 'uri'
 
 module Shelly
   
-  # For debugging purposes
-  DEBUG = {
-      :github => 'grimen/my_shell_scripts/install_geoip-city.sh/89ac494d95710c99a50e41d3a3b92ef8c474c88f',
-      :gist => '90101',
-      :pastie => '436580',
-      :pastebin => "m6758e75f"
-    }
-    
   VALID_SCRIPT_SOURCES = [:github, :gist, :pastie, :pastebin, :raw].freeze
   VALID_COMMANDS = {
+      :run => [],
       :create => [:config],
       :add => [:repo, :alias],
       :remove => [:repo, :alias],
@@ -50,22 +43,16 @@ module Shelly
   end
   
   def valid_command?(*args)
-    if args.size > 1
-      VALID_COMMANDS[args[0].to_sym].include?(args[1].to_sym)
+    if args[1]
+      VALID_COMMANDS[args[0].to_sym].empty? || VALID_COMMANDS[args[0].to_sym].include?(args[1].to_sym)
     else
       VALID_COMMANDS.keys.include?(args[0].to_sym)
     end
   end
   
-  def run!(args)
+  def run(args)
+    (args ||= []).collect! { |arg| arg.strip }
     
-    args ||= []
-    args.collect! { |arg| arg.strip }
-    
-    puts "============================================================================="
-    puts "  SHELLY"
-    puts "=============================================================================\n"
-      
     unless Shelly.internet_connection?
       puts "[shelly]: You are not connected to the Internet."
     else
@@ -76,10 +63,11 @@ module Shelly
         elsif valid_repo?(args[0])
           # TODO: Get repo
         end
-          
-        if Shelly.valid_command?(args[0])
+        
+        if args.size > 0
           if Shelly.valid_command?(args[0], args[1])
             case args[0].to_sym
+            when :run then Shelly.run!(args[1])
             when :create then Shelly.create!(args[1])
             when :add then Shelly.add!(args[1], args[2])
             when :remove then Shelly.remove!(args[1], args[2])
@@ -89,11 +77,25 @@ module Shelly
           else
             puts "[shelly]: FAIL: '#{args[1]}' is not a valid command for '#{args[0]}'. Try 'help' for help."
           end
-          
-        elsif valid_script_source?(script_source = args[0].split(':').first.to_sym)
-          # For debugging purposes
-          script_id = DEBUG[script_source]
-          
+        else
+          puts "[shelly]: FAIL: No valid command."
+        end
+        
+      else
+        puts "[shelly]: FAIL: No valid command."
+      end
+    end
+    
+  end
+  
+  def run!(what)
+    if what
+      begin
+        whats = what.split(':')
+        script_source = whats[0].to_sym
+        script_id = whats[1]
+        
+        if valid_script_source?(script_source)
           case script_source
           when :github then Shelly::ScriptSource::Scm::Github.new(script_id).run!
           when :gist then Shelly::ScriptSource::Plain::Gist.new(script_id).run!
@@ -104,38 +106,48 @@ module Shelly
         else
           Shelly::ScriptSource::Plain::Raw.new(script_id).run!
         end
-        
-      else
-        puts "[shelly]: FAIL: No valid command or script source specified."
+      rescue StandardError => e
+        puts "#{e}"
       end
+    else
+      puts "[shelly]: Invalid args for RUN."
     end
-    
-    print "\n"
-    # puts "[shelly]: END\n"
-    
   end
   
   def create!(what)
-    case what.to_sym
-    when :config then Shelly::Store::Base.create_root_config!
+    if what
+      case what.to_sym
+      when :config then Shelly::Store::Base.create_root_config!
+      end
+    else
+      puts "[shelly]: Invalid args for CREATE."
     end
   end
   
   def add!(to, what)
-    case to.to_sym
-    when :repo then Shelly::Store::Repo.add(what)
-    when :alias then Shelly::Store::Alias.add(what)
+    if to
+      case to.to_sym
+      when :repo then Shelly::Store::Repo.add(what)
+      when :alias then Shelly::Store::Alias.add(what)
+      end
+    else
+      puts "[shelly]: Invalid args for ADD."
     end
   end
   
   def remove!(from, what)
-    case from.to_sym
-    when :repo then Shelly::Store::Repo.remove(what)
-    when :alias then Shelly::Store::Alias.remove(what)
+    if from
+      case from.to_sym
+      when :repo then Shelly::Store::Repo.remove(what)
+      when :alias then Shelly::Store::Alias.remove(what)
+      end
+    else
+      puts "[shelly]: Invalid args for REMOVE."
     end
   end
   
   def list(from)
+    from ||= :all
     case from.to_sym
     when :repos then puts Shelly::Store::Repo.list
     when :aliases then puts Shelly::Store::Alias.list
@@ -145,7 +157,7 @@ module Shelly
   
   # TODO: Implement help instructions
   def help(from)
-    puts "No help available."
+    puts "[shelly]: No help available."
   end
   
   extend self
